jcbackend/juicycup_web_api/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
Pillow==9.3.0
django-cors-headers==4.2.0
psycopg2==2.9.5
django-otp==1.3.0
twilio==8.11.1


